requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.5.14
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
idna==3.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.314
langsmith==0.0.43
llama-index==0.8.45.post1
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.1
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.16.4
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
yarl==1.9.2
